Improve the Wikiprov interface
We don't have to worry about dropping in the library as much if we
clean up the interface. Here we create an interface that wraps all
the spargo needed complexity and then the Wikiprov complexity.

One complexity that is handled more rigorously now is the number of
threads that we choose to use. A default is set to 10 but the caller
can now set this within the package. A hard-coded value has been
removed.

In testing a few other issues were noticed like losing the ability to
parse properties and return provenance correctly. This gave us the
opportunity to improve validation of IRIs.